refactor: removendo '/' no fim dos endpoints e organizando package validations
