+++ HotelAPI/src/main/java/br/com/totvs/hotel/validation/annotation/EnumValue.java
@@ -1,5 +1,6 @@
-package br.com.totvs.hotel.validation;
+package br.com.totvs.hotel.validation.annotation;
 
+import br.com.totvs.hotel.validation.validator.EnumValueValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
@@ -10,11 +11,11 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = ReservaValidator.class)
-public @interface Reserva {
+@Constraint(validatedBy = EnumValueValidator.class)
+public @interface EnumValue {
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
-    String message() default "data informada inválida";
-    String pattern() default "dd/MM/yyyy HH:mm";
+    String message() default "enum informado inválido";
+    Class<? extends Enum<?>> enumerator();
 
 }
