+++ HotelAPI/src/main/java/br/com/totvs/hotel/validation/annotation/Idade.java
@@ -1,5 +1,6 @@
-package br.com.totvs.hotel.validation;
+package br.com.totvs.hotel.validation.annotation;
 
+import br.com.totvs.hotel.validation.validator.IdadeValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
@@ -10,10 +11,13 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = CEPValidator.class)
-public @interface CEP {
+@Constraint(validatedBy = IdadeValidator.class)
+public @interface Idade {
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
-    String message() default "cep informado é inválido";
+    String message() default "idade informada inválida";
+    String pattern() default "dd/MM/yyyy";
+    int minimum() default 0;
+    int maximum() default 100;
 
 }
