+++ HotelAPI/src/main/java/br/com/totvs/hotel/validation/annotation/Reserva.java
@@ -1,5 +1,6 @@
-package br.com.totvs.hotel.validation;
+package br.com.totvs.hotel.validation.annotation;
 
+import br.com.totvs.hotel.validation.validator.ReservaValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 
@@ -10,13 +11,11 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = IdadeValidator.class)
-public @interface Idade {
+@Constraint(validatedBy = ReservaValidator.class)
+public @interface Reserva {
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
-    String message() default "idade informada inválida";
-    String pattern() default "dd/MM/yyyy";
-    int minimum() default 0;
-    int maximum() default 100;
+    String message() default "data informada inválida";
+    String pattern() default "dd/MM/yyyy HH:mm";
 
 }
